backend/pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "gunicorn>=23.0.0",
    "python-decouple>=3.8",
]
